Added support for minimal kexec logic based on routines already available in the kernel to help with this.
Looked into target /proc/kallsyms and identified the address for
some of the required functions which are not exposed for modules
by default like
	Disable the non boot cpus
	setup mm for reboot
	flush pages to mem

However unable to find init_mm directly (for now using show_pte
to get it indirectly only for print dump).

Added logic to include code to run on the new system has include
file to the kernel module. And wrote two simple looping logics.
One of them loops for ever the other loops for sometime and then
exits.

For some reason the logic in the inner most kexec logic doesn't
seem to trigger always properly (especially the cpu_reset logic).
Have to debug this further with my own uart prints to see till
where it goes. But for now I have used inline assembly for the
cpu_reset (which was basically just loading the passed address
to pc register).