This pushes the kernel module to the target using adb.
It also gets the running kernels symbol table and shows
the list of hidden kernel functions used in the program
as well as their current entry points in the target.

Note that these entry points should be same between what
is their in the target kernel as well as what is used in
the module code. If they don't match it will have to be
updated as required in the module code.