Added support for module parameters for 	Type of kexec required 	Base address of buffer to use for execution if Fixed mode Inturn the same is used in the logic as required.
Moved my sleep function into a seperate .c/.h file called gen_utils.

Added support for uart printing at the C code level thro a uart_utils
file. Note that this can be used only till the mmu is disabled, after
that the address to use should be the physical address and inturn this
is handled in the assembly code which executes after the switch/restart).

Also the virtual address got using ioremap for the uart base address
is printed, to try the same also in the assembler code after the switch
to confirm that the mmu switch off was successful.

FIXME: The code in assembler file to cross check the switch is currently
not added.

Added logic to check the return value of the disable_nonboot_cpu code

FIXME: Have to check

1. If the cpu on which we are running is found to be the CPU1 instead
of CPU0, then calling disable_nonboot_cpus and quiting so that the
module can be reloaded once again to run on CPU0 (now that CPU1 will be
stopped) is a useful strategy or not.

2. If getting the local processor spin lock at the beggining itself
independent of anything else helps or not. Also how it impacts
printks, disable_nonboot_cpu logic etc needs to be checked.

3. Looking at what is happending, either the cache is not getting
flushed properly or one of the above is affecting or the logic
already available in the kernel is not sound (i.e has bugs).