Wrt lbhkvc_k.c -------------------- Added support for additional module parameters to allow
	a. Run the code to switch to even before switching but just after
	the cache flush following the memcopy

	b. Add support for modifying the uart address in the code to run
	from physical (assumed to be) to the virtual address allocated
	using ioremap. This is to verify if MMU switch off is taking effect
	or not.

NOTE: It is assumed that the last 4 bytes of the code to run contains the
address of the uart. And PatchUart1 will overwrite it with the virtual addr
allocated for specified uart.

Added support for Beagleboard(currently thro emulation) in addition to Nooktab.
In turn the address for the hidden functions in kernel for beagle is picked
from its system map file in the kernel compile tree/path.

Updated the module to use the new ioctl mechanism instead of the older/original
ioctl mechanism of linux kernel.

Updated code to call disable_nonboot_cpus only if it has been assigned a addr,
this is required in beagleboard because it is NOT SMP so this func is not
available in the first place or so will be the case in any NON SMP setup.

Wrt uart_utils
-----------------
Added code to print E on uart, when ever the uart flush takes a lot of time to
flush itself. This is for debug purpose. Inturn also updated the loop counts
to be #defines.

Added support for multiple devices by allowing the Physical address to be
defined in the header file based on DEVICE_?????? define. Currently Nooktab
(uart1-ttyO0) and Beagleboard (uart3-ttyO2) supported.

The uartPort base address variable is exported in the header, this is to
support the mpPatchUart1 option of the kernel module.