Makefile ------------ As part of adding support for multiple devices to be supported from same code A variable called DEVICE is expected to be assigned as required and inturn the same will be passed as #define to the source code during its compilation.
bloop3.S
------------
Added support for beagle's uart3 base address.
FIXME: Currently the base address has to be manually switched when moving from
one target device to another.

hkvc-make.sh
--------------
Added support for compiling for multiple devices.
For devices where I can compile and run the kernel myself, the System.map file
is used as the base to determine the hidden function addresses. However for
devices where someone else (the vendor) has decided the kernel to run, even
thou source access is available, the /proc/kallsyms from device is used as the
base for getting addresses.

The semi-automated logic to cross check the used addresses of the hidden funcs
match the target runtime addresses or not thro manual/visual cross check of
addresses has ben moved from install to this make shell script with a new arg
called check to this shell script (other than clean and no arg (which means
compile)).

Also the DEVICE variable is passed to the make command. Which inturn is passed
thro CFLAGS by the Makefile to the actual files.

install.sh
--------------
As explained above hidden func address dump moved out of this

The file being pushed to target updated to take account of the device name
appended to the kernel module file name.