hkvc-make.sh --------------- Added path for beagle android in office machine
hkvc-qemu_sd-img-utils.h
---------------------------

Added sudo for commands requiring root permissions. This was done so that one
doesn't have to give/enable ALL ALL sudo permission to the user/developer
account on the machine. Instead selective sudo permission can be given for
commands used in the script with sudo like losetup kpartx mount mkfs.????

Added a new special create command which doesn't run when all command is given.
This helps create a SD Image and partition and format it as required.

Split the previous mount into a new prepare and mount

Also updated mount command to mount with developer uid passed, so that the
developer can modify the mounted partitions.

Added a new ANDPATH variable, so that one doesn't have to edit it in multiple
locations.