Makefile ------------
Added the -march option for armv7-a to allow the use of DSB and ISB instr
in the asm files to be run.

Replaced the hard coded cross compiler name with a CROSS_COMPILE variable.

Added a possibility for verbose compile of kernel module

Added the new probe1 asm file

hkvc-qemu_sd-img-utils
--------------------------

Added a new boot.scr command which creates a boot.scr file, which will be
automatically run by the u-boot.bin provided it is configured for the same.
Which inturn is the case by default in u-boot. The same gets copied to the
first partition (vfat) in SD image by the copy command.

The all command will no longer trigger the run command. Instead a new command
called runall is required to trigger run along with the other sd image related
commands. This allows the image to be prepared/updated in one terminal and
inturn the same to be executed by qemu in a different terminal.

The check for default loop device available is not run/executed if the command
passed is run. Because losetup requires sudo/root permision while running
qemu doesn't require root permission. Thus avoiding the need to authenticate
with sudo during run command.

The run command now passes additional arguments to the qemu-system-arm command
related to no graphics as well as redirection of serial uarts into unix/linux
pty devices (accessible thro /dev/pts/[1|2|????).

FIXME: There seems to be a bug or improper handling of the serial option by
qemu, the order in which it assigns the serial port to ptys seems random and
also only 1 uart could be triggered by default. As far as only single uart
being available, not sure if it is due to the uart not being configured by
the linux kernel/user space.

lbhkvc_k.c and probe1.S
--------------------------
Have used DSB and ISB instructions as required, So that any issues due to
dual issue/pipeline and or out of order execution pipeline in Cortex-A9
core can be resolved by forcing a barrier for data operations and
instruction operations as required. Hopefully this was what was causing
random behaviour after / during the switch to the new code.

Also found that proc_fin and setup_mm_for_reboot don't disable MMU, so I
added explicit logic to switch off the MMU in the new asm file probe1.S

Also the value of the SCTRL of CP15 before and after the MMU switch is saved
into r4, r7(temp) and r8 respectively.

NOTE: If the asm files are not having a CAPTIAL S extension they get deleted
by the kernel related make clean used by module. So while creating any asm
file remember to end with a .S instead of .s extension.

Readme
---------
Added a comment on the default value of Arm CoProc CP15 SCTLR register which
gives details about MMU, Cache, etc is enabled or disabled, plus plus...