ASM Files for New pristine environment -----------------------------------------
Had a stupid, simple and serious bug in most of my .S files which among
many other more serious and convoluted issues was not allowing my
minimal kexec logic to succeed.

I had forgotten that in arm you require to explicitly tell arithmatic
and logical operations that they have to update the flag. Because I
had forgotten this, even if my linboot logic had worked, I wouldn't
have found it out because the prints supposed to come in the new
pristine environment wouldn't have got printed, as the code would
have got locked into a stupid loop before the actual print/write to
uart data register.

However there are issues plaguing linboot that are more complex and
convoluted than this, but even this is a critical bug in one sense.