lbhkvc_k.c --------------- Renamed the MMU disable logic enabling #define.
Added spinlock_irqsave to the last stage function block. As proc_fin doesn't
disable irqs.

Moved by sleep function to a later stage in the flow, so that I can see the
address of the allocated buffer, so that I can set breakpoints or snoop the
memory before further action if required.