lbhkvc_k.c ------------ Added a print to get the hkvc_minimal_kexec function address
TODO: Have to print the address of the buffer with in the kernel
module which containing code to be executed. And maybe just read
thro it or do some minimal operation involving it once so that
it gets loaded into memory and then use it to directly copy to
the destination buffer from with in my code in hkvc_minimal_kexec
rather than using memcopy outside.